+++ app/src/main/java/com/criteo/testapp/PrivacyConsentSelectorActivity.java
@@ -32,7 +32,6 @@ public class PrivacyConsentSelectorActivity extends AppCompatActivity {
   private static final String TAG = PrivacyConsentSelectorActivity.class.getSimpleName();
 
   private EditText ccpaIabEditText;
-  private EditText ccpaMopubEditText;
   private CheckBox ccpaBinaryValue;
   private Button saveBtn;
 
@@ -44,9 +43,6 @@ public class PrivacyConsentSelectorActivity extends AppCompatActivity {
     ccpaIabEditText = findViewById(R.id.editCcpaIabString);
     ccpaIabEditText.setText(getIabConsentString());
 
-    ccpaMopubEditText = findViewById(R.id.editCcpaMopubString);
-    ccpaMopubEditText.setText(getMopubString());
-
     ccpaBinaryValue = findViewById(R.id.checkboxCcpaBinary);
     ccpaBinaryValue.setChecked(getBinaryValue());
 
@@ -54,9 +50,6 @@ public class PrivacyConsentSelectorActivity extends AppCompatActivity {
 
     saveBtn.setOnClickListener(v -> {
       saveIabConsentString();
-
-      saveMopubConsentString();
-
       setBinaryValue();
     });
   }
@@ -81,13 +74,6 @@ public class PrivacyConsentSelectorActivity extends AppCompatActivity {
     return defaultSharedPreferences.getString("IABUSPrivacy_String", "");
   }
 
-  private String getMopubString() {
-    SharedPreferences defaultSharedPreferences = PreferenceManager
-        .getDefaultSharedPreferences(this);
-
-    return defaultSharedPreferences.getString("MoPubConsent_String", "");
-  }
-
   private void saveIabConsentString() {
     Log.d(TAG, "Storing IAB String: " + ccpaIabEditText.getText().toString());
 
@@ -96,17 +82,6 @@ public class PrivacyConsentSelectorActivity extends AppCompatActivity {
 
     Editor edit = defaultSharedPreferences.edit();
     edit.putString("IABUSPrivacy_String", ccpaIabEditText.getText().toString());
-    edit.apply();
-  }
-
-  private void saveMopubConsentString() {
-    Log.d(TAG, "Storing Mopub String: " + ccpaMopubEditText.getText().toString());
-
-    SharedPreferences defaultSharedPreferences = PreferenceManager
-        .getDefaultSharedPreferences(this);
-
-    Editor edit = defaultSharedPreferences.edit();
-    edit.putString("MoPubConsent_String", ccpaMopubEditText.getText().toString());
     edit.apply();
   }
 }