+++ test-utils/src/main/java/com/criteo/publisher/TestAdUnits.java
@@ -88,10 +88,6 @@ public class TestAdUnits {
   public static final RewardedAdUnit REWARDED_PREPROD = new RewardedAdUnit(
       "/140800857/Endeavour_RewardedVideo");
 
-  public static final String MOPUB_MEDIATION_BANNER_ADUNIT_ID = "b5acf501d2354859941b13030d2d848a";
-  public static final String MOPUB_MEDIATION_INTERSTITIAL_ADUNIT_ID = "86c36b6223ce4730acf52323de3baa93";
-  public static final String MOPUB_MEDIATION_NATIVE_ADUNIT_ID = "a298abc2fdf744cf898791831509cc38";
-
   public static final String ADMOB_MEDIATION_BANNER_ADUNIT_ID = "/140800857/Endeavour_320x50";
   public static final String ADMOB_MEDIATION_INTERSTITIAL_ADUNIT_ID = "/140800857/Endeavour_320x480";
   public static final String ADMOB_MEDIATION_NATIVE_ADUNIT_ID = "/140800857/Endeavour_Native";