+++ test-utils/src/main/java/com/criteo/publisher/network/CdbMock.kt
@@ -29,9 +29,6 @@ import com.criteo.publisher.TestAdUnits.INTERSTITIAL
 import com.criteo.publisher.TestAdUnits.INTERSTITIAL_PREPROD
 import com.criteo.publisher.TestAdUnits.INTERSTITIAL_VIDEO
 import com.criteo.publisher.TestAdUnits.INTERSTITIAL_VIDEO_PREPROD
-import com.criteo.publisher.TestAdUnits.MOPUB_MEDIATION_BANNER_ADUNIT_ID
-import com.criteo.publisher.TestAdUnits.MOPUB_MEDIATION_INTERSTITIAL_ADUNIT_ID
-import com.criteo.publisher.TestAdUnits.MOPUB_MEDIATION_NATIVE_ADUNIT_ID
 import com.criteo.publisher.TestAdUnits.NATIVE
 import com.criteo.publisher.TestAdUnits.NATIVE_PREPROD
 import com.criteo.publisher.TestAdUnits.REWARDED
@@ -252,8 +249,6 @@ class CdbMock(private val jsonSerializer: JsonSerializer) {
         bannerAdUnit in listOf(BANNER_320_50, BANNER_320_480, BANNER_320_50_PREPROD) ||
             interstitialAdUnit in listOf(INTERSTITIAL, INTERSTITIAL_PREPROD) ||
             placementId in listOf(
-                MOPUB_MEDIATION_BANNER_ADUNIT_ID,
-                MOPUB_MEDIATION_INTERSTITIAL_ADUNIT_ID,
                 ADMOB_MEDIATION_BANNER_ADUNIT_ID,
                 ADMOB_MEDIATION_INTERSTITIAL_ADUNIT_ID)
         -> {
@@ -308,7 +303,7 @@ class CdbMock(private val jsonSerializer: JsonSerializer) {
       """.trimIndent()
         }
         nativeAdUnit in listOf(NATIVE, NATIVE_PREPROD) ||
-            placementId in listOf(MOPUB_MEDIATION_NATIVE_ADUNIT_ID, ADMOB_MEDIATION_NATIVE_ADUNIT_ID) -> {
+            placementId in listOf(ADMOB_MEDIATION_NATIVE_ADUNIT_ID) -> {
           """
         {
           "impId": "$impressionId",